[CAN-78] fix: 글자수세기 반영되게 수정
